move propertypanels into their own package
