+++ meerk40t/gui/propertypanels/consoleproperty.py
@@ -1,7 +1,7 @@
 import wx
 
-from .icons import icons8_comments_50
-from .mwindow import MWindow
+from ..icons import icons8_comments_50
+from ..mwindow import MWindow
 
 _ = wx.GetTranslation
