+++ meerk40t/gui/propertypanels/groupproperties.py
@@ -1,8 +1,8 @@
 import wx
 
-from ..svgelements import SVG_ATTR_ID
-from .icons import icons8_group_objects_50
-from .mwindow import MWindow
+from ...svgelements import SVG_ATTR_ID
+from ..icons import icons8_group_objects_50
+from ..mwindow import MWindow
 
 _ = wx.GetTranslation
