+++ meerk40t/gui/propertypanels/imageproperty.py
@@ -1,8 +1,8 @@
 import wx
 
-from ..svgelements import Matrix
-from .icons import icons8_image_50
-from .mwindow import MWindow
+from ...svgelements import Matrix
+from ..icons import icons8_image_50
+from ..mwindow import MWindow
 
 _ = wx.GetTranslation
