+++ meerk40t/gui/propertypanels/operationpropertymain.py
@@ -2,8 +2,8 @@ import wx
 
 from meerk40t.kernel import signal_listener
 
-from ..svgelements import Color
-from .laserrender import swizzlecolor
+from ...svgelements import Color
+from ..laserrender import swizzlecolor
 
 _ = wx.GetTranslation
